+++ apps/e2e/tests/backend/endpoints/api/v1/projects.test.ts
@@ -83,7 +83,7 @@ it("creates and updates the basic project information of a project", async ({ ex
           "credential_enabled": true,
           "domains": [],
           "email_config": { "type": "shared" },
-          "email_theme": "default-light",
+          "email_theme": "<stripped UUID>",
           "enabled_oauth_providers": [],
           "magic_link_enabled": false,
           "oauth_account_merge_strategy": "link_method",
@@ -130,7 +130,7 @@ it("updates the basic project configuration", async ({ expect }) => {
           "credential_enabled": false,
           "domains": [],
           "email_config": { "type": "shared" },
-          "email_theme": "default-light",
+          "email_theme": "<stripped UUID>",
           "enabled_oauth_providers": [],
           "magic_link_enabled": true,
           "oauth_account_merge_strategy": "link_method",
@@ -182,7 +182,7 @@ it("updates the project domains configuration", async ({ expect }) => {
             },
           ],
           "email_config": { "type": "shared" },
-          "email_theme": "default-light",
+          "email_theme": "<stripped UUID>",
           "enabled_oauth_providers": [],
           "magic_link_enabled": false,
           "oauth_account_merge_strategy": "link_method",
@@ -240,7 +240,7 @@ it("updates the project domains configuration", async ({ expect }) => {
             },
           ],
           "email_config": { "type": "shared" },
-          "email_theme": "default-light",
+          "email_theme": "<stripped UUID>",
           "enabled_oauth_providers": [],
           "magic_link_enabled": false,
           "oauth_account_merge_strategy": "link_method",
@@ -292,7 +292,7 @@ it("should allow insecure HTTP connections if insecureHttp is true", async ({ ex
             },
           ],
           "email_config": { "type": "shared" },
-          "email_theme": "default-light",
+          "email_theme": "<stripped UUID>",
           "enabled_oauth_providers": [],
           "magic_link_enabled": false,
           "oauth_account_merge_strategy": "link_method",
@@ -387,7 +387,7 @@ it("updates the project email configuration", async ({ expect }) => {
             "type": "standard",
             "username": "test username",
           },
-          "email_theme": "default-light",
+          "email_theme": "<stripped UUID>",
           "enabled_oauth_providers": [],
           "magic_link_enabled": false,
           "oauth_account_merge_strategy": "link_method",
@@ -444,7 +444,7 @@ it("updates the project email configuration", async ({ expect }) => {
             "type": "standard",
             "username": "test username2",
           },
-          "email_theme": "default-light",
+          "email_theme": "<stripped UUID>",
           "enabled_oauth_providers": [],
           "magic_link_enabled": false,
           "oauth_account_merge_strategy": "link_method",
@@ -487,7 +487,7 @@ it("updates the project email configuration", async ({ expect }) => {
           "credential_enabled": true,
           "domains": [],
           "email_config": { "type": "shared" },
-          "email_theme": "default-light",
+          "email_theme": "<stripped UUID>",
           "enabled_oauth_providers": [],
           "magic_link_enabled": false,
           "oauth_account_merge_strategy": "link_method",
@@ -530,7 +530,7 @@ it("updates the project email configuration", async ({ expect }) => {
           "credential_enabled": true,
           "domains": [],
           "email_config": { "type": "shared" },
-          "email_theme": "default-light",
+          "email_theme": "<stripped UUID>",
           "enabled_oauth_providers": [],
           "magic_link_enabled": false,
           "oauth_account_merge_strategy": "link_method",
@@ -587,7 +587,7 @@ it("updates the project email configuration", async ({ expect }) => {
             "type": "standard",
             "username": "control group",
           },
-          "email_theme": "default-light",
+          "email_theme": "<stripped UUID>",
           "enabled_oauth_providers": [],
           "magic_link_enabled": false,
           "oauth_account_merge_strategy": "link_method",
@@ -751,13 +751,14 @@ it("updates the project oauth configuration", async ({ expect }) => {
           "credential_enabled": true,
           "domains": [],
           "email_config": { "type": "shared" },
-          "email_theme": "default-light",
+          "email_theme": "<stripped UUID>",
           "enabled_oauth_providers": [{ "id": "google" }],
           "magic_link_enabled": false,
           "oauth_account_merge_strategy": "link_method",
           "oauth_providers": [
             {
               "id": "google",
+              "provider_config_id": "google",
               "type": "shared",
             },
           ],
@@ -800,13 +801,14 @@ it("updates the project oauth configuration", async ({ expect }) => {
           "credential_enabled": true,
           "domains": [],
           "email_config": { "type": "shared" },
-          "email_theme": "default-light",
+          "email_theme": "<stripped UUID>",
           "enabled_oauth_providers": [{ "id": "google" }],
           "magic_link_enabled": false,
           "oauth_account_merge_strategy": "link_method",
           "oauth_providers": [
             {
               "id": "google",
+              "provider_config_id": "google",
               "type": "shared",
             },
           ],
@@ -851,7 +853,7 @@ it("updates the project oauth configuration", async ({ expect }) => {
           "credential_enabled": true,
           "domains": [],
           "email_config": { "type": "shared" },
-          "email_theme": "default-light",
+          "email_theme": "<stripped UUID>",
           "enabled_oauth_providers": [{ "id": "google" }],
           "magic_link_enabled": false,
           "oauth_account_merge_strategy": "link_method",
@@ -860,6 +862,7 @@ it("updates the project oauth configuration", async ({ expect }) => {
               "client_id": "client_id",
               "client_secret": "client_secret",
               "id": "google",
+              "provider_config_id": "google",
               "type": "standard",
             },
           ],
@@ -901,13 +904,14 @@ it("updates the project oauth configuration", async ({ expect }) => {
           "credential_enabled": true,
           "domains": [],
           "email_config": { "type": "shared" },
-          "email_theme": "default-light",
+          "email_theme": "<stripped UUID>",
           "enabled_oauth_providers": [{ "id": "spotify" }],
           "magic_link_enabled": false,
           "oauth_account_merge_strategy": "link_method",
           "oauth_providers": [
             {
               "id": "spotify",
+              "provider_config_id": "spotify",
               "type": "shared",
             },
           ],
@@ -956,7 +960,7 @@ it("updates the project oauth configuration", async ({ expect }) => {
           "credential_enabled": true,
           "domains": [],
           "email_config": { "type": "shared" },
-          "email_theme": "default-light",
+          "email_theme": "<stripped UUID>",
           "enabled_oauth_providers": [
             { "id": "google" },
             { "id": "spotify" },
@@ -966,10 +970,12 @@ it("updates the project oauth configuration", async ({ expect }) => {
           "oauth_providers": [
             {
               "id": "google",
+              "provider_config_id": "google",
               "type": "shared",
             },
             {
               "id": "spotify",
+              "provider_config_id": "spotify",
               "type": "shared",
             },
           ],
@@ -1018,7 +1024,7 @@ it("updates the project oauth configuration", async ({ expect }) => {
           "credential_enabled": true,
           "domains": [],
           "email_config": { "type": "shared" },
-          "email_theme": "default-light",
+          "email_theme": "<stripped UUID>",
           "enabled_oauth_providers": [
             { "id": "google" },
             { "id": "spotify" },
@@ -1028,10 +1034,12 @@ it("updates the project oauth configuration", async ({ expect }) => {
           "oauth_providers": [
             {
               "id": "google",
+              "provider_config_id": "google",
               "type": "shared",
             },
             {
               "id": "spotify",
+              "provider_config_id": "spotify",
               "type": "shared",
             },
           ],
@@ -1320,50 +1328,18 @@ it("should increment and decrement userCount when a user is added to a project",
       magic_link_enabled: true,
     }
   });
-  const initialProjectResponse = await niceBackendFetch("/api/v1/internal/projects/current", { accessType: "admin" });
+  const initialProjectResponse = await niceBackendFetch("/api/v1/internal/metrics", { accessType: "admin" });
   expect(initialProjectResponse.status).toBe(200);
+  expect(initialProjectResponse.body.total_users).toBe(0);
 
 
   // Create a new user in the project
   await Auth.Password.signUpWithEmail();
 
   // Check that the userCount has been incremented
-  const updatedProjectResponse = await niceBackendFetch("/api/v1/internal/projects/current", { accessType: "admin" });
+  const updatedProjectResponse = await niceBackendFetch("/api/v1/internal/metrics", { accessType: "admin" });
   expect(updatedProjectResponse.status).toBe(200);
-  expect(updatedProjectResponse).toMatchInlineSnapshot(`
-    NiceResponse {
-      "status": 200,
-      "body": {
-        "config": {
-          "allow_localhost": true,
-          "allow_team_api_keys": false,
-          "allow_user_api_keys": false,
-          "client_team_creation_enabled": false,
-          "client_user_deletion_enabled": false,
-          "create_team_on_sign_up": false,
-          "credential_enabled": true,
-          "domains": [],
-          "email_config": { "type": "shared" },
-          "email_theme": "default-light",
-          "enabled_oauth_providers": [],
-          "magic_link_enabled": true,
-          "oauth_account_merge_strategy": "link_method",
-          "oauth_providers": [],
-          "passkey_enabled": false,
-          "sign_up_enabled": true,
-          "team_creator_default_permissions": [{ "id": "team_admin" }],
-          "team_member_default_permissions": [{ "id": "team_member" }],
-          "user_default_permissions": [],
-        },
-        "created_at_millis": <stripped field 'created_at_millis'>,
-        "description": "",
-        "display_name": "New Project",
-        "id": "<stripped UUID>",
-        "is_production_mode": false,
-      },
-      "headers": Headers { <some fields may have been hidden> },
-    }
-  `);
+  expect(updatedProjectResponse.body.total_users).toBe(1);
 
   // Delete the user
   const deleteRes = await niceBackendFetch("/api/v1/users/me", {
@@ -1373,6 +1349,8 @@ it("should increment and decrement userCount when a user is added to a project",
   expect(deleteRes.status).toBe(200);
 
   // Check that the userCount has been decremented
-  const finalProjectResponse = await niceBackendFetch("/api/v1/internal/projects/current", { accessType: "admin" });
+  const finalProjectResponse = await niceBackendFetch("/api/v1/internal/metrics", { accessType: "admin" });
   expect(finalProjectResponse.status).toBe(200);
+  expect(finalProjectResponse.body.total_users).toBe(0);
+
 });