+++ packages/stack-ui/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @stackframe/stack-ui
 
+## 2.8.23
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.23
+
+## 2.8.22
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.22
+
+## 2.8.21
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.21
+
+## 2.8.20
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.20
+
 ## 2.8.19
 
 ### Patch Changes